Remove 404ed link, unwrap text
The link to the schema 404s, and I can't find such a document anywhere in the repo. Also, all of the lines of text were broken with carrige returns, which Markdown was rendering with `<br>` tags, leading to unfortunate text wrapping and oddly ragged edges, so I removed all single carriage returns and replaced them with spaces.